Move projections to server folder
