+++ x-pack/legacy/plugins/apm/common/projections/util/merge_projection/index.ts
