+++ x-pack/legacy/plugins/apm/public/components/app/ServiceOverview/index.tsx
@@ -17,7 +17,7 @@ import { ServiceList } from './ServiceList';
 import { useUrlParams } from '../../../hooks/useUrlParams';
 import { useTrackPageview } from '../../../../../infra/public';
 import { useCore } from '../../../hooks/useCore';
-import { PROJECTION } from '../../../projections/typings';
+import { PROJECTION } from '../../../../common/projections/typings';
 import { LocalUIFilters } from '../../shared/LocalUIFilters';
 
 const initalData = {
