+++ x-pack/legacy/plugins/apm/server/lib/errors/get_error_groups.ts
@@ -15,8 +15,8 @@ import {
 import { PromiseReturnType } from '../../../typings/common';
 import { APMError } from '../../../typings/es_schemas/ui/APMError';
 import { Setup } from '../helpers/setup_request';
-import { getErrorGroupsProjection } from '../../../public/projections/errors';
-import { mergeProjection } from '../../../public/projections/util/merge_projection';
+import { getErrorGroupsProjection } from '../../../common/projections/errors';
+import { mergeProjection } from '../../../common/projections/util/merge_projection';
 
 export type ErrorGroupListAPIResponse = PromiseReturnType<
   typeof getErrorGroups
